added file utils methods as separate utility
